fix the projectController

+++ app/Http/Controllers/Admin/ProjectController.php
@@ -17,7 +17,7 @@ class ProjectController extends Controller
             'title'=> 'required|string|min:5|max:100',
             'type_id' =>'required|integer|exists:types,id',
             'url_image'=> 'required|url|max:200',
-            'image'=> 'image|max:1024',
+            'image'=> 'nullable|image|max:1024',
             'repo'=> 'required|string|min:5|max:100',
             'languages'=> 'required|string|min:5|max:100',
             'description'=> 'required|string|min:5',
@@ -38,14 +38,10 @@ class ProjectController extends Controller
     public function index()
     {
        $projects = Project::paginate(10);
+       
        return view('admin.projects.index', compact('projects'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
         $types = Type::all();
@@ -62,24 +58,21 @@ class ProjectController extends Controller
     public function store(Request $request)
     {
 
-        // richiedere($data) e validare i dati del form
         $request->validate($this->validations, $this->validation_messages);
+
         $data = $request->all();
 
-        // salvare l'immagine caricata
-        // prendere il percorso dell'immagine appena salvata
-        $imagePath = Storage::put("uploads", $data["image"]);
- 
-        // salvare i dati se corretti
+        $imagePath = Storage::put('uploads', $data['image']);
+
         $newProject = new Project();
         $newProject->title          = $data['title'];
         $newProject->slug           = Project::slugger($data["title"]);
         $newProject->type_id        = $data['type_id'];
         $newProject->url_image      = $data['url_image'];
-        $newProject->image      = $imagePath;
         $newProject->repo           = $data['repo'];
         $newProject->languages      = $data['languages'];
         $newProject->description    = $data['description'];
+        $newProject->image          = $imagePath;
         $newProject->save();
 
         $newProject->technologies()->sync($data["technologies"] ?? []);
@@ -132,18 +125,17 @@ class ProjectController extends Controller
         $project = Project::where("slug", $slug)->firstOrFail();
 
         $request->validate($this->validations, $this->validation_messages);
+        
         // richiedere($data) e validare i dati del form
         $data = $request->all();
 
-        // salvare l'immagine nuova
-        $imagePath = Storage::put("uploads", $data["image"]);
-
-        // eliminare l'immagine vecchia
-        if($project->image) {
-            Storage::delete($project->image);
+        if($data["image"]) {
+            $imagePath = Storage::put("uploads", $data["image"]);
+            if($project->image) {
+                Storage::delete($project->image);
+            }
+            $project->image = $imagePath;
         }
-
-        $project->image = $imagePath;
 
         // salvare i dati se corretti
         $project->title = $data['title'];
@@ -166,48 +158,19 @@ class ProjectController extends Controller
      * @param  \App\Models\Project  $project
      * @return \Illuminate\Http\Response
      */
-    public function delete($slug)
+    public function destroy(Project $project)
     {
+        if ($project->image) {
+            Storage::delete($project->image);
+        }
 
-        $project = Project::where("slug", $slug)->firstOrFail();
-
+        // dissociare tutti i tag
         $project->technologies()->detach();
+
+        // eliminare il portfolio
 
         $project->delete();
+
         return to_route('admin.projects.index')->with('delete_success', $project);
-    }
-
-    public function restore($id)
-    {
-        Project::withTrashed()->where('id', $id)->restore();
-
-        $project = Project::find($id);
-
-        return to_route('admin.projects.trashed')->with('restore_success', $project);
-    }
-
-    public function cancel($id)
-    {
-        Project::withTrashed()->where('id', $id)->restore();
-
-        $project = Project::find($id);
-
-        return to_route('admin.project.index')->with('cancel_success', $project);
-    }
-
-    public function trashed()
-    {
-        $trashedProjects = Project::onlyTrashed()->paginate(3);
-
-        return view('admin.projects.trashed', compact('trashedProjects'));
-    }
-
-    public function harddelete($id)
-    {
-        $project = Project::withTrashed()->find($id);
-        $project->technologies()->detach();
-        
-
-        return to_route('admin.projects.trashed')->with('delete_success', $project);
     }
 }
